Entrypoint adjustments
