+++ enterprise/frontend/src/metabase-enterprise/embedding_iframe_sdk/types/embed.ts
@@ -1,13 +1,13 @@
+import type { MetabaseError } from "embedding-sdk-bundle/errors";
+import type { MetabaseAuthMethod } from "embedding-sdk-bundle/types";
+import type { MetabaseEmbeddingSessionToken } from "embedding-sdk-bundle/types/refresh-token";
 import type {
   CollectionBrowserListColumns,
   EmbeddingEntityType,
   EntityTypeFilterKeys,
   MetabaseTheme,
   SqlParameterValues,
-} from "embedding-sdk-bundle";
-import type { MetabaseError } from "embedding-sdk-bundle/errors";
-import type { MetabaseAuthMethod } from "embedding-sdk-bundle/types";
-import type { MetabaseEmbeddingSessionToken } from "embedding-sdk-bundle/types/refresh-token";
+} from "embedding-sdk-package";
 import type { ParameterValues } from "metabase/embedding-sdk/types/dashboard";
 import type { CollectionId } from "metabase-types/api";
